# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Invoice Generator API
  description: |
    SaaS-ready invoice generator with multi-tenancy, multi-currency,
    jurisdiction-aware tax rules, recurring billing, and PDF generation.

    ## Architecture
    - Hexagonal Architecture (Ports & Adapters)
    - CQRS (Command Query Responsibility Segregation)
    - Domain Events + Outbox Pattern
    - Row-Level Security for tenant isolation

    ## Authentication
    All protected endpoints require a Bearer JWT token.
    Obtain a token via `POST /api/v1/auth/login`.

    ## Idempotency
    POST endpoints that create or modify state support the `Idempotency-Key` header.
    Provide a unique key (e.g., UUID) to safely retry requests without side effects.
    Cached responses are returned for duplicate keys within a 24-hour window.
  version: 1.0.0
  contact:
    name: Santiago SSAA
    url: https://github.com/santiagossaa
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.invoicegen.example.com
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new tenant
      description: Creates a new tenant organization and admin user.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Tenant and user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      description: Authenticates user and returns JWT token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token obtained via /api/v1/auth/login
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: invalid request body
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RegisterRequest:
      type: object
      required: [tenant_name, email, password]
      properties:
        tenant_name:
          type: string
          example: ACME Corp
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expires_at:
          type: string
          format: date-time
        tenant_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid